models: store domains_count on Instance
Rather than computing this for a specific use of instance, use a hook to
keep the value up to date.

Domains_count has a odd property that it is the same for all instance's
this pub service supports. Although, this might change in the future
when domain blocks are added.