activitypub: fix unvalidated deletes
There's a problem with deletes. If an actor is removed from a remote
server, it seems that the fediverse broadcasts their demise widely. But,
if they're deleted, and we don't have a copy of their record locally, we
cannot validate their signature. But, if we don't have their signature,
there's nothing to delete.

Update the verification logic to move it lower in the processor logic,
then we can see if we have the actor locally first, and if so, verify
their request for deletion.